bgreydon/fix int output
Fixed incorrect format output of integers. Most likely, all cases have been tested on Linux.

+++ src/s21_string/s21_sprintf/s21_sprintf.h
@@ -14,7 +14,7 @@
 
 int s21_sprintf_(char *str, const char *format, va_list args);
 
-void int_number_to_char(char **str, unsigned long long int number,
+void int_number_to_char(char **str, unsigned long number,
                         format_info *info);
 
 void real_number_to_char(char **str, double number, format_info *info);
